convert kernel to float

+++ Cxx11/transpose.cl
@@ -6,14 +6,14 @@
 // - https://github.com/sschaetz/nvidia-opencl-examples/blob/master/OpenCL/src/oclTranspose/transpose.cl
 //
 __kernel void transpose(const int order,
-                        __global double * a,
-                        __global double * b)
+                        __global float * a,
+                        __global float * b)
 {
     const int i = get_global_id(0);
     const int j = get_global_id(1);
 
     if ((i<order) && (j<order)) {
         b[i*order+j] += a[j*order+i];
-        a[j*order+i] += 1.0;
+        a[j*order+i] += 1.0f;
     }
 }
